ai_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiohttp==3.11.10
Pillow==11.0.0
numpy==2.1.3
ultralytics==8.3.40
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
torchvision==0.20.1+cpu
